Printf error, where long strings to printf break the code
Long strings to printf  end up in EXC_BAD_ADDRESS or the program prints
some of the strings held in other string buffers.  It may be that the char
pointer loaded.moduleName is being overwritten.  However, the debugger
suggests that this is not the case.  I can't figure out the issue.