New structure, the UseNode for the useList.
With the UseNode structure, we can save the external address with the symbol itself.  This way we can update the external address when we look at the instruction.  In addition, we can check to see if the external address remains 0.  Although we have no reason to do so at the moment.